requirements.txt
langchain_openai
langchain_community
langchain_core
langchain
openai
streamlit
huggingface_hub
python_dotenv
PyPDF2
faiss-cpu
langchain_google_genai
google.generativeai
langchain_google_genai